refactor: adjustment package and database
