+++ src/app.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const routes = require('./routes');
+require('./database');
 
 const app = express();
