+++ src/services/UserService.js
@@ -1,4 +1,4 @@
-const { User } = require('../models');
+const { User } = require('../models/User');
 
 class UserService {
     async createUser(userData) {
